requirements.txt
pandas<1.2.0
xlrd<2.0
streamlit>=1.8.1
plotly==4.14.3
Pillow==8.1.2
protobuf~=3.19.0
xlsxWriter==3.0.3
scipy==1.7.2
numpy==1.21
requests==2.28.0
xlwt==1.3
PySimpleGUI==4.60.3